srihealth-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
passlib[bcrypt]
pyjwt
torch
torchvision
pillow
